:memo: refactor: 조직 투두 schemas 기술 (#51)
