+++ src/main/java/com/notitime/noffice/request/AnnouncementCreateRequest.java
@@ -3,7 +3,6 @@ package com.notitime.noffice.request;
 import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.notitime.noffice.response.TaskCreateRequest;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.media.Schema.RequiredMode;
 import java.util.List;
