+++ src/main/java/com/notitime/noffice/request/AnnouncementUpdateRequest.java
@@ -1,6 +1,5 @@
 package com.notitime.noffice.request;
 
-import com.notitime.noffice.response.TaskCreateRequest;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.media.Schema.RequiredMode;
 import java.util.List;
