+++ src/main/java/com/notitime/noffice/request/TaskCreateRequest.java
@@ -1,0 +1,10 @@
+package com.notitime.noffice.request;
+
+import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+
+public record TaskCreateRequest(
+		@Schema(description = "투두 항목", example = "노피스 API 제작", requiredMode = REQUIRED)
+		String content) {
+}
