--- src/main/java/com/notitime/noffice/response/TaskCreateRequest.java
@@ -1,4 +1,0 @@
-package com.notitime.noffice.response;
-
-public record TaskCreateRequest(String content) {
-}
